🚚 Move constants.py to src directory
